+++ backend/src/main/java/com/tienda/backend/dto/CompraRequestDTO.java
@@ -3,33 +3,33 @@ package com.tienda.backend.dto;
 import java.util.List;
 
 public class CompraRequestDTO {
-    private Long personaId; // Se solicita temporalmente ya que no hay login en Sprint 2
-    private Long metodoPagoId;
+    private Integer personaId; // Se solicita temporalmente ya que no hay login en Sprint 2
+    private Integer metodoPagoId;
     private List<CompraDetalleDTO> detalles;
 
     // Constructores
     public CompraRequestDTO() {}
 
-    public CompraRequestDTO(Long personaId, Long metodoPagoId, List<CompraDetalleDTO> detalles) {
+    public CompraRequestDTO(Integer personaId, Integer metodoPagoId, List<CompraDetalleDTO> detalles) {
         this.personaId = personaId;
         this.metodoPagoId = metodoPagoId;
         this.detalles = detalles;
     }
 
     // Getters y Setters
-    public Long getPersonaId() {
+    public Integer getPersonaId() {
         return personaId;
     }
 
-    public void setPersonaId(Long personaId) {
+    public void setPersonaId(Integer personaId) {
         this.personaId = personaId;
     }
 
-    public Long getMetodoPagoId() {
+    public Integer getMetodoPagoId() {
         return metodoPagoId;
     }
 
-    public void setMetodoPagoId(Long metodoPagoId) {
+    public void setMetodoPagoId(Integer metodoPagoId) {
         this.metodoPagoId = metodoPagoId;
     }
